# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation SPN",
        "description": "Ikhwan Computer",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "afd.zik@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/spnbackend/public/api/klasifikasi/list": {
            "get": {
                "tags": [
                    "Klasifikasi Surat"
                ],
                "summary": "List Klasifikasi Surat",
                "description": "List Klasifikasi Surat",
                "operationId": "klasifikasi",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/spnbackend/public/api/suratKeluar/list": {
            "get": {
                "tags": [
                    "Surat Keluar"
                ],
                "summary": "List Surat Keluar",
                "description": "API List Surat Keluar",
                "operationId": "suratKeluar",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "take",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/spnbackend/public/api/suratKeluar/view/{id}": {
            "get": {
                "tags": [
                    "Surat Keluar"
                ],
                "summary": "Detail surat Keluar by ID",
                "description": "Detail Surat Keluar By ID",
                "operationId": "suratKeluar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/spnbackend/public/api/suratMasuk/list": {
            "get": {
                "tags": [
                    "Surat Masuk"
                ],
                "summary": "List Surat Masuk",
                "description": "API List Surat Masuk",
                "operationId": "suratMasuk",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "take",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/spnbackend/public/api/suratMasuk/view/{id}": {
            "get": {
                "tags": [
                    "Surat Masuk"
                ],
                "summary": "Detail surat masuk by ID",
                "description": "Detail Surat Masuk By ID",
                "operationId": "suratMasuk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}